baseline testing setup for karpatkey role mgmt
