+++ packages/sdk/test/karpatkey/transactions/.gitignore
@@ -1,0 +1,1 @@
+*New.json
