+++ packages/sdk/test/karpatkey/transactions/ltdManageGnosisChain.json
@@ -1,0 +1,106 @@
+{
+  "permissions": { "targets": [] },
+  "success": [
+    "0x3c7aac5ff890ca8a32a836f5476bbfc37de86b27a0ec9afdb16fa3bb34fa7bf4",
+    "0x60a270abd5116c05366586d03dba36a0bae4bac2166f61052392c6a67989bce0",
+    "0xdc6e5ebce78e7812d8a68f6f79fb9686a1eb9f13a6621244ebbb3c1d50bd496b",
+    "0xade58c2226d3a2ee9fdd1440f6556b3b96fd6912cc587d9e169961ec7c9d0d60",
+    "0x35599ce2bf54bc1401c203a197d75db50dd49b505885ce24693f0ad6ff1d15ab",
+    "0xad0f2d18e1920f83ae8e403ba7383c27d3aed45bdbb3ce19be1822a4e3b19867",
+    "0x026aff7528960acae03f5574f60d49a93f769e188fb89d4ce0cff1608caba394",
+    "0xa41843202189efb9c0f8bed8e1fcb29020ea996dc8c6bc6b87079ce3eb89869a",
+    "0xc80c51da7771331388b44eceaae7c4a46dee118943b63c4c1959c180803d0f39",
+    "0x6237db76083290d6fb0f2a8f9d9b9d33c6e65762c31a70dc18c6a170dc8a72c5",
+    "0x3e5561627b6d2653021e37095e9d92d51c98d3b8844db49e4a66b60a85248f03",
+    "0x81902404883cf1df742199ee2d0b9f139f1514c72f35478db861ce6e232280a7",
+    "0xf3bc817f85f9b2ee932e54ceac7ef0b8132ef3c6d64a8105c87f65fe453de1b8",
+    "0x63f1c1a6ce3d04fba3685f9e973a4dd78c3b6a8a0af2d9800cd278d26b6b6ca2",
+    "0xf29cbdf665db2995a90f8324626776d6feb8bc6c04eee159435b84a5764026c2",
+    "0x6d3a4bc4fa59fd7d94bddf2a4bb36d91e6af8a8641db0292b0cf96e3d4d54b4e",
+    "0xd8762aae0521f3845bc52e85fb95218e8de9ed7ecb598acfa7cb7088acf62d18",
+    "0x368d1a02d063ace40f2a4f400faf85ad387ecb21c003e5d1cc0fd9896ac9d70d",
+    "0xbbf094f4b7042a705cd840813f7cf9f92db84d699f8ddcec95acc3d2e9e9f0ef",
+    "0x79e7582877f4cc7d3d21842c8d75261f4e0f4de0722e1a9c9f26995af818ec16",
+    "0x5f231aaf0b582a8a94ca9b02b3d880218810c0cf7217aa608dd718af3f8a6da2",
+    "0xaf8672b9d218b24fa64a3ebbac0fcf60a524d2625f09b412407d1b5cae617fc1",
+    "0xd4d1544cd9de668129a8f25c5c9c43b88e30c9283684ce34f735ff53ae1b75df",
+    "0xc09e8b71938bd8db12d0ca5bb98126c60a5a2026c6132cc888a9138b12857db8",
+    "0x76c82d4801c9da5554ad058c187a33748e54fb58fea5139eb8bab700cc0c296c",
+    "0x2357d00177fce59fcb0259c63761e3c6333f06788aa50f3f3fbb09bb8b7163b0",
+    "0x6975fc50fe8046bda180b74dd3cbdbddd3157c73cd398e13f815f9ee4ff31c4f",
+    "0xe0703e8f24cc8c179cb64072a63da73ab324f68ac1340116a2338315dc0630d3",
+    "0x013ed7c632934af3f2ac5b158e8e8b8470dce5da9297b758d885c020adf2e6c9",
+    "0x2a2aaf172b54271c9181ff7bc0609767010cd49cc948ec7d456dd5ac2ada1441",
+    "0x08156df9aa5295d6c2aa33a1e4a531716d4d7709495077b57bc142a2fb87d8f6",
+    "0x8cd0c7442790214e5ab34fc8f361d09e0e7a53c192e245acf8aea1d1106bdc0f",
+    "0xe92788114e6db3629d18f392faf9480a2c55f6e0b478149075199ac5b4bdcf1d",
+    "0x021e9b4c9a59ff9ec0d51e564d2a98a97f24236cd4fed34766e756143e58ed66",
+    "0xbb81b0d2d65e5d47f57e95f3231cd8134a151f64a782d5357c6b268b63c0d761",
+    "0xa5d567a6eea3c23f2b848d099ab3fa7c066f3a89e226f83ac241bc7f8d39c38e",
+    "0x4fdd2190a703a6bf03aa1fd63cc02d38b6d645cdea824ce0f4806e24de9bbe46",
+    "0xa2581e6ffe04a57a3ca6791d1f3b8690d896c6cac5b196492bd4bc76ab6c751a",
+    "0xd162fd31509a5732a35bbf0f7d0446a308807412f34a4aab493a1374aaf1798a",
+    "0x21dd48586b313e1b5ce450d50b02c83a30b42e2383d641ebf92a49f39a00974c",
+    "0x92078ddd2a44bb9656a1683fbec1c7704b0a24ac902ac590a2af97c92b2a8caf",
+    "0x6ca92cc12ab458c1eefc094c2c3d28d09f96b76ed0dd77d281f189f821b594ab",
+    "0xb7f9de69c0f90941b85e328dee8d41a2e0f2ee17f6b74dc81c02b7809b3f6c68",
+    "0xb982b22828f7f5010fb9eadf6e900b694a48f936265719e8d8ca6ed77f3043c0",
+    "0x80102ceede438c3d0f905cad7478a81179b327811f40f67c8ec3b6d4c6d48bce",
+    "0x9fb134b3e76be000404c05f18c315e8fd6f23c08d06c45f6b514ca81db99af13",
+    "0x926f1edc26a439995aaf02519476ede9fcba4f8a9138e9d274dc3be0884b85f8",
+    "0x3ea3509abe69aad7cbd144824f567d9cc5293786bb06f51a20a2d3b3a6026aa6",
+    "0x7add278d9e81e40a4bd8a1f26ed0e7216c27ec8777d032e0382e12a5c5c2e945",
+    "0x7f40b390e2b42ea84d33d7d6727135575195a7c789a30fe8bd3716c2f056a2d4"
+  ],
+  "fail": [
+    "0x5beb987553ce2d8cd2d033f82a46b3bc5f76552f51e942446ff080c65533469e",
+    "0xdc191fec6b1d9dc2c534b28da06822440c2fe06533f0043a7def01a7b2d13567",
+    "0x0ead19b2cfe3ad38245faf3e6168d6ee61980c2a8a331abc834a471bee105c34",
+    "0x9aa034e45830d002c266725e634cd4660ae61be9d8c80011726f124e6932f251",
+    "0xfc577f6a18b79888e2c4299608e332b46cde0598a093d3c3d6db37067dd01c14",
+    "0x51efaf3e4e93648ad469f6c6affce24820c0d0dbd99f2c9c2ec0d8b7f1ad2335",
+    "0x8f06ecd412d9e982b335efea1e5818c53314ab06903d70e5b69f0b2300f0a35d",
+    "0xcb4bf28b42cdcfca272a27594c07d7a6d589d76317781d52b3426b221571cf85",
+    "0x8f28884b516673bb872efc0ee46b8e4a4b721d69f6368c24254b3d38de6ad349",
+    "0x02bfad9755901ba39bfcb246549d01e2df6dfed6095a347c3432c117ee6db7b3",
+    "0xb9135836aaa41c6a67e3ef09ca1816318e10bf6887f59e8a26d5ad23ecb4136d",
+    "0x01cb777c1c2dfe63a1afe12af7f478b47e019066496f7404ff42e33372f71b24",
+    "0x0c37bdfc001e78c19c85d51c69e49059cb916000f23c80e0efc2d0edcc430212",
+    "0xd58a25882a13b946bb1ad034785ed10bd5e7645674480f23f95a903f7ff58d61",
+    "0x92602cda56b84b0bb04f4adaffb012cc59d28bfb5bcb6d6c258070abb9172557",
+    "0xba9b28496e9930863c4b3618bfe5d518ba35719135fd02ea1eaffa1ce8a0ed07",
+    "0xbc6fd9405b9f47145ef4cd0a35b3323543ccdee967fea73bc2338b576aa74640",
+    "0x6bf9b4a05c32d21ef99895471dea175e9af8492bc57e4f488326b6bbd568b4be",
+    "0x45fb8a25181f401149259b47e8596934743bd8502e0f281da189307f326cf770",
+    "0xcbeea61e5d3f8a3e87b7701e62b9ffde9efa314123328c799c1f7df49b0d2609",
+    "0x6b29f73a303dbeb1a16126e187c598b655d0a9b2a00b88b4bf78e34e9d88bc51",
+    "0x79d5f0cdcdde14b1b18948b5f9c1893549a73b048ce26b300a680458775addd2",
+    "0x4e38f7758008b0751f1423da79d47e913e9eec6241a3ab1f31dfdba511595245",
+    "0x83978ccf49ce262d49b256318dc674ba7cae30ceed61306e7b8d3d067fe2637a",
+    "0xb5a5bb3b764f41d4f11c991806299d1b6f204974da8c2eeef811db50217791f5",
+    "0xd7c839910ca075e4fc62bd5c23cc5339c475a32e4182cc06ee39c7ce62370edb",
+    "0x7e71c0d30f62bf6980c7aafed3fde4355420248305107c20c89f9e4a58fb20ae",
+    "0x2138b716c0dfc441c0d0d0c5214d8cb44978040adcfa99929eada376f7f04cb1",
+    "0xb9a0702626b29823b001543ac86e1e289c3052fde756d8ca9e9756fce6ffb849",
+    "0x640170dfbdeb79567887d41fe78f61cfb6237d87f5b39acb1c24b1d3928b10ed",
+    "0x077af2631ae7cc0bfdd020852d1bb27cb4219c7fef842f565f6f7e34e6f801a4",
+    "0xaa9530b8f6e0e103ad598ea7c1756f0bdfe83577d18f050811b6f65756e58b70",
+    "0x276484e2eec44f1af9547994878c31f7cf17a65d5d07031495db02c503f93917",
+    "0xd6ea3d4d542886bb8b3d08c87bbb7b76fa458704c44ccc1c666df46fbd4bc8b4",
+    "0xfbafc05ad23a9cc66c4393ed4e3f06cc603a2313db8531a28a8473c8f5203dba",
+    "0x64c8635afa4763dbd1099cf341347240f02bf0ed44baf738dd242df55eb39b87",
+    "0x9a1241bd8c966bb2b2036236919b9dbe5ab16d1a631c018e83c55bf765f59f8c",
+    "0xab342937b53b94348f7949dd2cb746d00fd2de4002b8bee98f4925fad9553fcc",
+    "0x9fbbdfd8e2d420e482749e3c25c9c511f7e6813cba9c8a40c399dc8a1d05f349",
+    "0x321059dde1bc9ea95dda6f03f731d9c11c937070fead9baf73065cf7ec677341",
+    "0x3beab1eb9a52b63814faf47302742c2938967cb54848a7734e1c0d3ca2b38c32",
+    "0x4df807bd42efe312901a529df097ffc216f062338ac713165861e81fd086cc9d",
+    "0x6cec403a9d2884f774b6255ce9791faafeabf066361d949b6fadaba6ae31aac8",
+    "0xb55bd2528e5a992a5128ab5e0b9890bca7e6e758a54b327b8eb33461b1c7cc76",
+    "0x86352bd6e648c09dac52ca2e5929a6a188638acb7bbade734da9bbe205a4007b",
+    "0xd2e11482a3837ea959fba447988087f520c85ccefcf89d9f6067c47e04b2fa1e",
+    "0xcf4596e6754cb51ee0294ec9f50a7897d1f9426d9ed4b5d3b092744bff993168",
+    "0x425c0e06ff036e3aaac3a1ae9a6b0896ef93bc09f2ad1a33a054857c21c6daa6",
+    "0x59a07879e7c6d6733f91d570c6fa97c60884993563e92fce20afa5771ae6f9f5"
+  ]
+}
